- f2fs: fix missing cp.reason is CP_SYNC
The cp.reason should be assigned.
Otherwise, if cp.reason has obsolete data like CP_DISCARD, it causes kernel
panic like:
 Unable to handle kernel paging request at virtual address 32d1b444
 pgd = da044000
 [32d1b444] *pgd=00000000
 Internal error: Oops: 5 [#1] PREEMPT SMP ARM
 Modules linked in: wlan(O) touchx [last unloaded: wlan]
 CPU: 2 Tainted: G O (3.4.42-g97affd0-00015-g5946620 #1)
 PC is at add_discard_addrs+0x4c/0x294
 LR is at flush_sit_entries+0x5fc/0x724
...
 (add_discard_addrs+0x4c/0x294) from [] (flush_sit_entries+0x5fc/0x724)
 (flush_sit_entries+0x5fc/0x724) from [] (write_checkpoint+0x290/0xb34)
 (write_checkpoint+0x290/0xb34) from [] (f2fs_gc+0xd0/0x3b4)
 (f2fs_gc+0xd0/0x3b4) from [] (f2fs_write_begin+0xcc/0x54c)

This is because add_discard_addrs assumes that cpc->trim_start and
cpc->trim_end should be correctly assigned when cpc->reason is CP_DISCARD.
But, in this case, it didn't, which causes the panic in the following line.

	if (test_bit(cpc->trim_start, dirty_i->dirty_segmap[PRE])) {
                          ----------

This is occurred in stable-v3.X only, because two patches were changed this
line in dev-v3.X but one of them was not merged in stable-v3.X.

Change-Id: Ida68336c979910e13d84489b306c3b8f96cbec9a
Signed-off-by: Jaegeuk Kim <jaegeuk@motorola.com>
Signed-off-by: Yahya Wessam <yahyawessam2002@gmail.com>